making a new branch for the development.

+++ ntl/linked_ptr.hxx
@@ -1,15 +1,15 @@
 /**\file*********************************************************************
  *                                                                     \brief
- *  Linked pointer  
+ *  Linked pointer
  *
  ****************************************************************************
  */
-
 #ifndef NTL__LINKED_PTR
 #define NTL__LINKED_PTR
- 
-#include "stlx/algorithm.hxx"
+#pragma once
+
 #include "linked_list.hxx"
+#include "stdlib.hxx"
 
 namespace ntl {
 
@@ -17,7 +17,6 @@ template<typename T>
 struct linked_ptr
 {
     typedef linked_ptr  this_type;
-    typedef T(this_type::*unspecified_bool_type)();
 
     typedef T         value_type;
     typedef T*        pointer;
@@ -42,9 +41,9 @@ struct linked_ptr
 
     /// unlink
     ~linked_ptr()
-    { 
+    {
       /// unless this is not the last one
-      //if ( links.prev != links.next )
+      if ( links.prev != links.next )
         links.unlink(links.prev, links.next);
     }
 
@@ -54,7 +53,7 @@ struct linked_ptr
       long c = 0;
       const_double_linked_iterator i = &links;
       do if ( containing_adaptor<this_type, double_linked, &this_type::links>(i)->ptr )
-          ++c; 
+          ++c;
       while ( --i != &links );
       return c;
     }
@@ -65,10 +64,10 @@ struct linked_ptr
       return links.prev == links.next && ptr;
     }
 
-    operator unspecified_bool_type() const
-    { 
+    operator explicit_bool_type() const
+    {
       // ptr != 0 forces a few bloat instructions
-      return brute_cast<unspecified_bool_type>(ptr);
+      return explicit_bool(ptr);
     }
 
     void swap(this_type & r)
